[Silver III] Title: 2×n 타일링, Time: 0 ms, Memory: 2024 KB -BaekjoonHub

+++ 백준/Silver/11726. 2×n 타일링/2×n 타일링.cc
@@ -6,7 +6,7 @@
 using namespace std;
 
 int N;
-long long arr[1001];
+int arr[1001];
 
 int main() {
